requirements/requirements-core.txt
jinja2
PyYAML
rich
typer
types-PyYAML
